Fix all 402 findings from fresh exhaustive codebase review
Addresses every finding from a full multi-agent review across all 16
subsystems and all 5 backends (CPU, CUDA, ROCm, OneAPI, Vulkan), with
full correct fixes only — no workarounds, no CPU fallbacks for GPU backends.

Severity breakdown:
- 10 critical (4 regressions + 6 bugs), incl. SvdBackward output-slot
  routing, RPN graph severing, and the systemic MKL ILP64/lp64 build
  mismatch root cause (also unblocked det/LDL/sparse).
- 106 high across autograd, all backends, nn, distributed, models,
  jit/lazy, io, python.
- 126 medium + 160 low across every subsystem.

Highlights:
- Autograd: SVD higher-order skew form + non-square projector terms,
  PermuteBackward neg-dim, fusion consumer edges, vmap graph severing,
  scatter contiguity guards, jvp half-widening.
- Backends: half/bf16 widen-in-kernel (no CPU fallback), stride-aware
  kernels, OOB/bounds checks, grid-stride tail coverage, allocator
  invalid-free/leak fixes, GLSL shader correctness (cross_entropy,
  bicubic, sparse strides, atomics), int4 alignment guards.
- Security/robustness: overflow-safe untrusted-file parsing (lite/jit
  serialization), histogram/image bounds, TFRecord LE CRC.
- nn/models: grad_fn-preserving forwards (RoPE/MoE/attention), forward
  saved-stats, scheduler/optimizer state fixes, QuantizedConvTranspose2d
  delegation, per-channel observer neg-axis normalization.

All 5 backend libraries compile clean; affected CPU test suites pass.
Corrected AdagradLearningRateDecay test to match the (PyTorch-consistent)
decay the optimizer actually applies (the old expectation only passed due
to a get_lr off-by-one bug).

+++ include/tenzor/io/stb/stb_image_write.h
@@ -584,7 +584,7 @@ static int stbi_write_tga_core(stbi__write_context *s, int x, int y, int comp, v
                   }
                } else {
                   for (k = i + 2; k < x && len < 128; ++k) {
-                     if (!memcmp(begin, row + k * comp, comp)) {
+                     if (!memcmp(begin, row + (size_t)k * (size_t)comp, comp)) {
                         ++len;
                      } else {
                         break;
